rename: UserPoint, PointHistory 계층 변경 (dto -> entity)
