+++ src/main/java/io/hhplus/tdd/database/PointHistoryTable.java
@@ -1,8 +1,8 @@
 package io.hhplus.tdd.database;
 
 
-import io.hhplus.tdd.point.dto.PointHistory;
-import io.hhplus.tdd.point.dto.TransactionType;
+import io.hhplus.tdd.point.entity.PointHistory;
+import io.hhplus.tdd.point.entity.TransactionType;
 import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
