+++ src/main/java/io/hhplus/tdd/point/entity/PointHistory.java
@@ -1,4 +1,4 @@
-package io.hhplus.tdd.point.dto;
+package io.hhplus.tdd.point.entity;
 
 public record PointHistory(
         long id,
